Refactoring - move common Spreadsheet code to extensions class
