+++ src/Sfa.Tl.Matching.Application.UnitTests/FileWriter/OpportunityPipelineReport/When_OpportunityReport_Writer_Is_Called_To_Create_Spreadsheet_With_Referrals_Only.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using DocumentFormat.OpenXml.Packaging;
 using DocumentFormat.OpenXml.Spreadsheet;
 using FluentAssertions;
+using Sfa.Tl.Matching.Application.Extensions;
 using Sfa.Tl.Matching.Application.FileWriter.Opportunity;
 using Sfa.Tl.Matching.Application.UnitTests.FileWriter.OpportunityPipelineReport.Builders;
 using Xunit;
@@ -12,7 +13,6 @@ namespace Sfa.Tl.Matching.Application.UnitTests.FileWriter.OpportunityPipelineRe
     public class When_OpportunityReport_Writer_Is_Called_To_Create_Spreadsheet_With_Referrals_Only
     {
         private readonly byte[] _result;
-        private readonly OpportunityPipelineReportWriter _reportWriter;
 
         public When_OpportunityReport_Writer_Is_Called_To_Create_Spreadsheet_With_Referrals_Only()
         {
@@ -20,8 +20,8 @@ namespace Sfa.Tl.Matching.Application.UnitTests.FileWriter.OpportunityPipelineRe
                     .AddReferralItem()
                     .Build();
 
-            _reportWriter = new OpportunityPipelineReportWriter();
-            _result = _reportWriter.WriteReport(dto);
+            var reportWriter = new OpportunityPipelineReportWriter();
+            _result = reportWriter.WriteReport(dto);
         }
 
         [Fact]
@@ -52,7 +52,7 @@ namespace Sfa.Tl.Matching.Application.UnitTests.FileWriter.OpportunityPipelineRe
             {
                 using (var spreadSheet = SpreadsheetDocument.Open(stream, false))
                 {
-                    var sheetData = _reportWriter.GetSheetData(spreadSheet, 0);
+                    var sheetData = spreadSheet.GetSheetData(0);
                     sheetData.Should().NotBeNull();
 
                     var rows = sheetData.Descendants<Row>().ToList();
