+++ src/Sfa.Tl.Matching.Application.UnitTests/FileWriter/ProviderProximityReport/When_ProviderProximityReport_Writer_Is_Called_To_Create_Spreadsheet.cs
@@ -3,6 +3,7 @@ using System.Linq;
 using DocumentFormat.OpenXml.Packaging;
 using DocumentFormat.OpenXml.Spreadsheet;
 using FluentAssertions;
+using Sfa.Tl.Matching.Application.Extensions;
 using Sfa.Tl.Matching.Application.FileWriter.Provider;
 using Sfa.Tl.Matching.Application.UnitTests.FileWriter.ProviderProximityReport.Builders;
 using Xunit;
@@ -12,7 +13,6 @@ namespace Sfa.Tl.Matching.Application.UnitTests.FileWriter.ProviderProximityRepo
     public class When_ProviderProximityReport_Writer_Is_Called_To_Create_Spreadsheet
     {
         private readonly byte[] _result;
-        private readonly ProviderProximityReportWriter _reportWriter;
 
         public When_ProviderProximityReport_Writer_Is_Called_To_Create_Spreadsheet()
         {
@@ -20,8 +20,8 @@ namespace Sfa.Tl.Matching.Application.UnitTests.FileWriter.ProviderProximityRepo
                     .AddProvider()
                     .Build();
 
-            _reportWriter = new ProviderProximityReportWriter();
-            _result = _reportWriter.WriteReport(dto);
+            var reportWriter = new ProviderProximityReportWriter();
+            _result = reportWriter.WriteReport(dto);
         }
 
         [Fact]
@@ -52,7 +52,7 @@ namespace Sfa.Tl.Matching.Application.UnitTests.FileWriter.ProviderProximityRepo
             {
                 using (var spreadSheet = SpreadsheetDocument.Open(stream, false))
                 {
-                    var sheetData = _reportWriter.GetSheetData(spreadSheet, 0);
+                    var sheetData = spreadSheet.GetSheetData(0);
                     sheetData.Should().NotBeNull();
 
                     var rows = sheetData.Descendants<Row>().ToList();
