+++ src/Sfa.Tl.Matching.Application/FileReader/ExcelFileReader.cs
@@ -3,14 +3,12 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Reflection;
-using System.Text;
 using System.Threading.Tasks;
-using DocumentFormat.OpenXml;
 using DocumentFormat.OpenXml.Packaging;
-using DocumentFormat.OpenXml.Spreadsheet;
 using FluentValidation;
 using FluentValidation.Results;
 using Microsoft.Extensions.Logging;
+using Sfa.Tl.Matching.Application.Extensions;
 using Sfa.Tl.Matching.Application.Interfaces;
 using Sfa.Tl.Matching.Data.Interfaces;
 using Sfa.Tl.Matching.Domain.Models;
@@ -44,7 +42,7 @@ namespace Sfa.Tl.Matching.Application.FileReader
 
             using (var document = SpreadsheetDocument.Open(fileImportDto.FileDataStream, false))
             {
-                var rows = GetAllRows(document, fileImportDto.NumberOfHeaderRows).ToList();
+                var rows = document.GetAllRows(fileImportDto.NumberOfHeaderRows).ToList();
 
                 var stringTablePart = document.WorkbookPart.SharedStringTablePart;
 
@@ -59,9 +57,9 @@ namespace Sfa.Tl.Matching.Application.FileReader
                 {
                     foreach (var column in columnProperties)
                     {
-                        var cell = GetCellByIndex(column.Index, startIndex, row);
+                        var cell = row.GetCellByIndex(column.Index, startIndex);
 
-                        var cellValue = GetCellValue(stringTablePart, cell);
+                        var cellValue = stringTablePart.GetCellValue(cell);
 
                         column.ColumnInfo.SetValue(fileImportDto, cellValue.Trim());
                     }
@@ -96,77 +94,6 @@ namespace Sfa.Tl.Matching.Application.FileReader
             await _functionLogRepository.CreateManyAsync(validationErrors);
 
             return dtos;
-        }
-
-        private static IEnumerable<Row> GetAllRows(SpreadsheetDocument document, int? headerRowIndex)
-        {
-            var workbookPart = document.WorkbookPart;
-            var sheets = workbookPart.Workbook.GetFirstChild<Sheets>().Elements<Sheet>();
-            var relationshipId = sheets.First().Id.Value;
-            var worksheetPart = (WorksheetPart)document.WorkbookPart.GetPartById(relationshipId);
-            var workSheet = worksheetPart.Worksheet;
-            var sheetData = workSheet.GetFirstChild<SheetData>();
-            var rows = sheetData.Descendants<Row>();
-
-            if (headerRowIndex.HasValue)
-                rows = rows.Skip(headerRowIndex.Value);
-
-            return rows;
-        }
-
-        private static Cell GetCellByIndex(int cellIndex, int rowIndex, OpenXmlElement row)
-        {
-            var cellReference = GetCellReferenceByIndex(cellIndex, rowIndex);
-            return row.Descendants<Cell>().FirstOrDefault(cell => cell.CellReference == cellReference);
-        }
-
-        private static string GetCellValue(SharedStringTablePart stringTablePart, CellType cell)
-        {
-            var cellValue = string.Empty;
-
-            if (cell == null) return cellValue;
-
-            if (cell.DataType != null)
-            {
-                switch (cell.DataType.Value)
-                {
-                    case CellValues.SharedString:
-                        cellValue = stringTablePart.SharedStringTable.ChildElements[int.Parse(cell.CellValue.InnerXml)].InnerText;
-                        break;
-                    case CellValues.InlineString:
-                        cellValue = cell.InnerText;
-                        break;
-                    case CellValues.Boolean:
-                    case CellValues.Number:
-                    case CellValues.Error:
-                    case CellValues.String:
-                    case CellValues.Date:
-                        break;
-                    default:
-                        throw new ArgumentOutOfRangeException();
-                }
-            }
-            else if (cell.CellValue != null)
-            {
-                cellValue = cell.CellValue.InnerXml;
-            }
-
-            return cellValue;
-        }
-
-        public static string GetCellReferenceByIndex(int col, int row)
-        {
-            col++;
-            var sb = new StringBuilder();
-            do
-            {
-                col--;
-                sb.Insert(0, (char)('A' + col % 26));
-                col /= 26;
-
-            } while (col > 0);
-            sb.Append(row + 1);
-            return sb.ToString();
         }
 
         private void LogErrorsAndWarnings(int rowIndex, ValidationResult validationResult, List<FunctionLog> validationErrors)
