+++ src/Sfa.Tl.Matching.Application/FileWriter/Opportunity/OpportunityPipelineReportWriter.cs
@@ -6,13 +6,15 @@ using DocumentFormat.OpenXml;
 using DocumentFormat.OpenXml.Packaging;
 using DocumentFormat.OpenXml.Spreadsheet;
 using Sfa.Tl.Matching.Application.Constants;
+using Sfa.Tl.Matching.Application.Extensions;
+using Sfa.Tl.Matching.Application.Interfaces;
 using Sfa.Tl.Matching.Models.Dto;
 
 namespace Sfa.Tl.Matching.Application.FileWriter.Opportunity
 {
-    public class OpportunityPipelineReportWriter : ExcelFileWriter<OpportunityReportDto>
+    public class OpportunityPipelineReportWriter : IFileWriter<OpportunityReportDto>
     {
-        public override byte[] WriteReport(OpportunityReportDto data)
+        public byte[] WriteReport(OpportunityReportDto data)
         {
             var assembly = Assembly.GetExecutingAssembly();
             var templateName = ApplicationConstants.PipelineOpportunitiesReportTemplate;
@@ -32,10 +34,10 @@ namespace Sfa.Tl.Matching.Application.FileWriter.Opportunity
 
                 using (var spreadSheet = SpreadsheetDocument.Open(stream, true))
                 {
-                    var referralsSheetData = GetSheetData(spreadSheet, 0);
+                    var referralsSheetData = spreadSheet.GetSheetData(0);
                     WriteReferralsToSheet(data, referralsSheetData);
 
-                    var provisionGapsheetData = GetSheetData(spreadSheet, 1);
+                    var provisionGapsheetData = spreadSheet.GetSheetData(1);
                     WriteProvisionGapsToSheet(data, provisionGapsheetData);
 
                     RemoveEmptySheets(spreadSheet, data);
@@ -56,17 +58,17 @@ namespace Sfa.Tl.Matching.Application.FileWriter.Opportunity
 
             if (data.ReferralItems.Count == 0)
             {
-                sheetIdsToRemove.Add(GetSheetId(spreadSheet, 0));
+                sheetIdsToRemove.Add(spreadSheet.GetSheetId(0));
             }
 
             if (data.ProvisionGapItems.Count == 0)
             {
-                sheetIdsToRemove.Add(GetSheetId(spreadSheet, 1));
+                sheetIdsToRemove.Add(spreadSheet.GetSheetId(1));
             }
 
             foreach (var sheetId in sheetIdsToRemove)
             {
-                DeleteSheet(spreadSheet, sheetId);
+                spreadSheet.DeleteSheet(sheetId);
             }
         }
 
@@ -78,22 +80,22 @@ namespace Sfa.Tl.Matching.Application.FileWriter.Opportunity
             {
                 var row = new Row { RowIndex = (uint)rowIndex };
 
-                row.AppendChild(CreateTextCell(1, rowIndex, referral.Workplace));
-                row.AppendChild(CreateTextCell(2, rowIndex, referral.JobRole));
+                row.AppendChild(SpreadsheetExtensions.CreateTextCell(1, rowIndex, referral.Workplace));
+                row.AppendChild(SpreadsheetExtensions.CreateTextCell(2, rowIndex, referral.JobRole));
 
                 row.AppendChild(int.TryParse(referral.PlacementsDetail, out var placements)
-                    ? CreateTextCell(3, rowIndex, placements)
-                    : CreateTextCell(3, rowIndex, referral.PlacementsDetail));
+                    ? SpreadsheetExtensions.CreateTextCell(3, rowIndex, placements)
+                    : SpreadsheetExtensions.CreateTextCell(3, rowIndex, referral.PlacementsDetail));
 
-                row.AppendChild(CreateTextCell(4, rowIndex, referral.ProviderNameForReport));
-                row.AppendChild(CreateTextCell(5, rowIndex, referral.ProviderVenueTownAndPostcode));
-                row.AppendChild(CreateTextCell(6, rowIndex, $"{referral.DistanceFromEmployer:#0.0} miles"));
-                row.AppendChild(CreateTextCell(7, rowIndex, referral.PrimaryContact));
-                row.AppendChild(CreateTextCell(8, rowIndex, referral.PrimaryContactEmail));
-                row.AppendChild(CreateTextCell(9, rowIndex, referral.PrimaryContactPhone));
-                row.AppendChild(CreateTextCell(10, rowIndex, referral.SecondaryContact));
-                row.AppendChild(CreateTextCell(11, rowIndex, referral.SecondaryContactEmail));
-                row.AppendChild(CreateTextCell(12, rowIndex, referral.SecondaryContactPhone));
+                row.AppendChild(SpreadsheetExtensions.CreateTextCell(4, rowIndex, referral.ProviderNameForReport));
+                row.AppendChild(SpreadsheetExtensions.CreateTextCell(5, rowIndex, referral.ProviderVenueTownAndPostcode));
+                row.AppendChild(SpreadsheetExtensions.CreateTextCell(6, rowIndex, $"{referral.DistanceFromEmployer:#0.0} miles"));
+                row.AppendChild(SpreadsheetExtensions.CreateTextCell(7, rowIndex, referral.PrimaryContact));
+                row.AppendChild(SpreadsheetExtensions.CreateTextCell(8, rowIndex, referral.PrimaryContactEmail));
+                row.AppendChild(SpreadsheetExtensions.CreateTextCell(9, rowIndex, referral.PrimaryContactPhone));
+                row.AppendChild(SpreadsheetExtensions.CreateTextCell(10, rowIndex, referral.SecondaryContact));
+                row.AppendChild(SpreadsheetExtensions.CreateTextCell(11, rowIndex, referral.SecondaryContactEmail));
+                row.AppendChild(SpreadsheetExtensions.CreateTextCell(12, rowIndex, referral.SecondaryContactPhone));
 
                 sheetData.AppendChild(row);
                 rowIndex++;
@@ -108,13 +110,13 @@ namespace Sfa.Tl.Matching.Application.FileWriter.Opportunity
             {
                 var row = new Row { RowIndex = (uint)rowIndex };
 
-                row.AppendChild(CreateTextCell(1, rowIndex, provisionGap.Workplace));
-                row.AppendChild(CreateTextCell(2, rowIndex, provisionGap.JobRole));
+                row.AppendChild(SpreadsheetExtensions.CreateTextCell(1, rowIndex, provisionGap.Workplace));
+                row.AppendChild(SpreadsheetExtensions.CreateTextCell(2, rowIndex, provisionGap.JobRole));
                 row.AppendChild(int.TryParse(provisionGap.PlacementsDetail, out var placements)
-                    ? CreateTextCell(3, rowIndex, placements)
-                    : CreateTextCell(3, rowIndex, provisionGap.PlacementsDetail));
+                    ? SpreadsheetExtensions.CreateTextCell(3, rowIndex, placements)
+                    : SpreadsheetExtensions.CreateTextCell(3, rowIndex, provisionGap.PlacementsDetail));
 
-                row.AppendChild(CreateTextCell(4, rowIndex, provisionGap.Reason));
+                row.AppendChild(SpreadsheetExtensions.CreateTextCell(4, rowIndex, provisionGap.Reason));
 
                 sheetData.AppendChild(row);
                 rowIndex++;
